Patch `pandas.core.dtypes.inference.is_dict_like`
Pandas thinks all non-atom K objects are dicts, which leads to issues
like #133. This commit provides a hacky solution to this problem by
overwriting the `__code__` attribute of the function Pandas uses for
checking if an object is dict-like. Patching the function through safer
means (e.g. with `unittest.mock`) won't work well because the function
is imported all over Pandas, and we would have to patch all of them.

Fixes #133